+++ src/gallery-sm-tree/js/plugins/tree-lazy.js
@@ -1,19 +1,23 @@
+/*jshint expr:true, maxlen:200, onevar:false */
+
 /**
-Provides `Plugin.Tree.Lazy`, a plugin for `Tree` that makes it easy to lazily
-load and populate the contents of tree nodes the first time they're opened.
+Provides `Plugin.Tree.Lazy`, a plugin for `Tree.Openable` that makes it easy to
+lazily load and populate the contents of tree nodes the first time they're
+opened.
 
 @module gallery-sm-tree
 @submodule gallery-sm-tree-lazy
 **/
 
 /**
-A plugin for `Tree` that makes it easy to lazily load and populate the contents
-of tree nodes the first time they're opened.
+A plugin for `Tree.Openable` that makes it easy to lazily load and populate the
+contents of tree nodes the first time they're opened.
 
 ### Example
 
-    YUI().use('jsonp', 'tree', 'tree-lazy', function (Y) {
-        var tree = new Y.Tree();
+    YUI().use('jsonp', 'gallery-sm-tree-openable', 'gallery-sm-tree-lazy', function (Y) {
+        var Tree = Y.Base.create('openableTree', Y.Tree, [Y.Tree.Openable]),
+            tree = new Tree();
 
         tree.plug(Y.Plugin.Tree.Lazy, {
 
@@ -99,6 +103,12 @@ Y.namespace('Plugin.Tree').Lazy = Y.Base.create('lazyTreePlugin', Y.Plugin.Base,
             this.load = config.load;
         }
 
+        // Make sure we've been plugged into a Tree that mixes in the
+        // Tree.Openable extension.
+        if (!this._host.openNode) {
+            Y.log("Plugin.Tree.Lazy was plugged into a Tree that doesn't mix in the Tree.Openable extension. This probably won't do you much good.", 'warn', 'gallery-sm-tree-lazy');
+        }
+
         this._published = {};
         this._attachEvents();
     },
@@ -109,13 +119,11 @@ Y.namespace('Plugin.Tree').Lazy = Y.Base.create('lazyTreePlugin', Y.Plugin.Base,
     },
 
     // -- Protected Methods ----------------------------------------------------
-
     _attachEvents: function () {
         this.onHostEvent('open', this._onOpen);
     },
 
     // -- Protected Event Handlers ---------------------------------------------
-
     _onOpen: function (e) {
         var node = e.node;
 
@@ -135,7 +143,6 @@ Y.namespace('Plugin.Tree').Lazy = Y.Base.create('lazyTreePlugin', Y.Plugin.Base,
     },
 
     // -- Default Event Handlers -----------------------------------------------
-
     _defLoadingFn: function (e) {
         var node = e.node,
             self = this;