+++ src/gallery-sm-tree/js/tree.js
@@ -1,3 +1,5 @@
+/*jshint expr:true, onevar:false */
+
 /**
 Provides the `Tree` class.
 
@@ -48,24 +50,6 @@ var Lang = Y.Lang,
     @preventable _defClearFn
     **/
     EVT_CLEAR = 'clear',
-
-    /**
-    Fired when a node is closed.
-
-    @event close
-    @param {Tree.Node} node Node being closed.
-    @preventable _defCloseFn
-    **/
-    EVT_CLOSE = 'close',
-
-    /**
-    Fired when a node is opened.
-
-    @event open
-    @param {Tree.Node} node Node being opened.
-    @preventable _defOpenFn
-    **/
-    EVT_OPEN = 'open',
 
     /**
     Fired when a node is removed from this Tree.
@@ -253,26 +237,6 @@ var Tree = Y.Base.create('tree', Y.Base, [], {
             defaultFn: this._defClearFn,
             silent   : options && options.silent
         });
-    },
-
-    /**
-    Closes the specified node if it isn't already closed.
-
-    @method closeNode
-    @param {Object} [options] Options.
-        @param {Boolean} [options.silent=false] If `true`, the `close` event
-            will be suppressed.
-    @chainable
-    **/
-    closeNode: function (node, options) {
-        if (node.canHaveChildren && node.isOpen()) {
-            this._fire(EVT_CLOSE, {node: node}, {
-                defaultFn: this._defCloseFn,
-                silent   : options && options.silent
-            });
-        }
-
-        return this;
     },
 
     /**
@@ -468,26 +432,6 @@ var Tree = Y.Base.create('tree', Y.Base, [], {
     },
 
     /**
-    Opens the specified node if it isn't already open.
-
-    @method openNode
-    @param {Object} [options] Options.
-        @param {Boolean} [options.silent=false] If `true`, the `open` event
-            will be suppressed.
-    @chainable
-    **/
-    openNode: function (node, options) {
-        if (node.canHaveChildren && !node.isOpen()) {
-            this._fire(EVT_OPEN, {node: node}, {
-                defaultFn: this._defOpenFn,
-                silent   : options && options.silent
-            });
-        }
-
-        return this;
-    },
-
-    /**
     Prepends a node or array of nodes at the beginning of the specified parent
     node.
 
@@ -553,21 +497,6 @@ var Tree = Y.Base.create('tree', Y.Base, [], {
     **/
     size: function () {
         return this.rootNode.size();
-    },
-
-    /**
-    Toggles the open/closed state of the specified node.
-
-    @method toggleNode
-    @param {Tree.Node} node Node to toggle.
-    @param {Object} [options] Options.
-        @param {Boolean} [options.silent=false] If `true`, events will be
-            suppressed.
-    @chainable
-    **/
-    toggleNode: function (node, options) {
-        return node.isOpen() ? this.closeNode(node, options) :
-            this.openNode(node, options);
     },
 
     /**
@@ -756,14 +685,6 @@ var Tree = Y.Base.create('tree', Y.Base, [], {
 
         this.rootNode = newRootNode;
         this.children = newRootNode.children;
-    },
-
-    _defCloseFn: function (e) {
-        delete e.node.state.open;
-    },
-
-    _defOpenFn: function (e) {
-        e.node.state.open = true;
     },
 
     _defRemoveFn: function (e) {