+++ src/gallery-sm-tree/HISTORY.md
@@ -1,4 +1,18 @@
 SmugMug Tree History
 ====================
 
-Nothing yet!
+## git
+
+* Arbitrary properties in the `config` object passed to `Tree.Node`'s
+  constructor are now mixed into the instance, allowing you to attach your own
+  metadata or methods to nodes. Arbitrary config properties cannot overwrite
+  properties that already exist on the node or its prototype.
+
+## 2013-01-09
+
+* Moved the open/close/toggle functionality into a new
+  `gallery-sm-tree-openable` extension.
+
+## 2012-12-26
+
+* Initial gallery release.